优化软 I2C 配置，添加纳秒延迟参数并调整相关函数实现

+++ espidf/main/soft_i2c.h
@@ -28,7 +28,7 @@ typedef struct {
     bool pullup_en;
     bool inited;
     uint32_t clk_speed;
-    uint8_t ns;
+    uint32_t ns;
 } soft_i2c_port_t;
